requirements.txt
# AutoGen and AI dependencies
autogen-agentchat>=0.2.0
autogen-ext[openai]>=0.2.0
openai>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web and API utilities
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0

# File and data handling
pydantic>=2.5.0
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Subprocess and system integration
psutil>=5.9.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.26.0